Fix test to match new debug output
Prior to this change, the debug format always printed 4 spaces and then
the index of the error in the chain. Previous changes made it so that in
the cause chain we indent each new line to line up with the first line.
Together these two changes mean that if the index is longer than 1
character the first line will become misaligned. This was fixed by using
a left pad to 4 characters so the : is always on the 5th character and
all lines line up consistently.