Workaround for IntelliJ plugin
For _some_ reason the `intellij-rust-native-helper` (brought in via the RUSTC_WRAPPER env var) used by the rust plugin results in an exit code of 0 from compiling probe.rs, even when running on the stable toolchain, when `--emit=metadata` is set. While to true fix it certainly downstream with the JetBrains folk, it doesn't seem as though it is actually required here, so may be more expedient to fix here.