Enable backtrace support for >= 1.65 stable compilers
The `std::backtrace` APIs relevant for this crate were stabilized in
1.65. This commit makes anyhow use those API if possible.
`std::backtrace` is preferred over the `backtrace` crate, so for 1.65+
compilers, enabling the `backtrace` crate feature does nothing
essentially.

Previously, when a nightly compiler was used, use of `std::backtrace`
as well as the Provider API was enabled. This commit separates those
two APIs such that `std::backtrace` is used for nightly OR 1.65+ stable
compilers; and the Provider API is used for nightly compilers only.